requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
flask>=2.3.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
werkzeug>=2.3.0
scikit-learn>=1.2.0
